refactor: error 파일들 수정
