+++ backend/src/main/java/com/khu/bbangting/error/ResourceNotFoundException.java
@@ -1,4 +1,4 @@
-package com.khu.bbangting.security.exception;
+package com.khu.bbangting.error;
 
 public class ResourceNotFoundException extends RuntimeException {
